[diffusion] move precision helper under runtime utils
Keep shared precision utilities with the rest of runtime utils and update all multimodal generation imports/tests to use the new module path.
