+++ python/sglang/multimodal_gen/runtime/pipelines/comfyui_flux_pipeline.py
@@ -21,9 +21,9 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages import (
     ComfyUILatentPreparationStage,
     DenoisingStage,
 )
-from sglang.multimodal_gen.runtime.precision import resolve_precision
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
+from sglang.multimodal_gen.runtime.utils.precision import resolve_precision
 
 logger = init_logger(__name__)
