+++ python/sglang/multimodal_gen/runtime/pipelines/comfyui_zimage_pipeline.py
@@ -36,9 +36,9 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages import (
     ComfyUILatentPreparationStage,
     DenoisingStage,
 )
-from sglang.multimodal_gen.runtime.precision import resolve_precision
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
+from sglang.multimodal_gen.runtime.utils.precision import resolve_precision
 from sglang.multimodal_gen.utils import set_mixed_precision_policy
 
 logger = init_logger(__name__)
