+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/causal_denoising.py
@@ -29,15 +29,15 @@ from sglang.multimodal_gen.runtime.platforms import (
     AttentionBackendEnum,
     current_platform,
 )
-from sglang.multimodal_gen.runtime.precision import (
-    autocast_enabled as precision_autocast_enabled,
-)
 from sglang.multimodal_gen.runtime.realtime.states import (
     RealtimeCausalDiTState,
     get_realtime_causal_dit_state,
 )
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
+from sglang.multimodal_gen.runtime.utils.precision import (
+    autocast_enabled as precision_autocast_enabled,
+)
 
 logger = init_logger(__name__)
