+++ python/sglang/multimodal_gen/runtime/utils/precision.py
